% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyData.R
\docType{package}
\name{buoydata-package}
\alias{buoydata}
\alias{buoydata-package}
\title{buoydata: Extract buoy data from ndbc}
\description{
Tools to aid in extraction of data from the nbdc website
}
\details{
\itemize{
\item Get_ functions to get data.
\item functions to manipulate the data, aggregate the data
}

To learn more about using \code{buoydata} click the index below
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NOAA-EDAB/buoydata}
  \item Report bugs at \url{https://github.com/NOAA-EDAB/buoydata/issues}
}

}
\author{
\strong{Maintainer}: Andy Beet \email{andrew.beet@noaa.gov}

}
\keyword{internal}
