% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_buoy_data.R
\name{combine_buoy_data}
\alias{combine_buoy_data}
\title{Combine annual buoy data files}
\usage{
combine_buoy_data(buoyid = "bzbm3", variable, inDir)
}
\arguments{
\item{buoyid}{Character vector. Vector of buoy id names}

\item{variable}{Character string. Name of the variable to pull}

\item{inDir}{Character string. Full path to where annual files are located}
}
\value{
dataframe (nx4)
\item{YEAR}{Year}
\item{MONTH}{Month}
\item{DAY}{Day}
\item{"Variable"}{The name of the variable passed to the function}
}
\description{
All files available for a buoy are concatenated over time. Caution: some variables have changed in name over time
}
