% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buoy_data.R
\name{get_buoy_data}
\alias{get_buoy_data}
\title{Gets buoy data from ERDDAP}
\usage{
get_buoy_data(buoyid, vars = NULL)
}
\arguments{
\item{buoyid}{Character. Name of buoys id. See \code{\link{buoy_data}} for list of buoys available}

\item{vars}{Character vector. Vector of variables to pull. Default = NULL, pull all available variables. If you select spcific variables the resulting data frame will be considerably smaller}
}
\value{
A data frame
}
\description{
Pulls data from ERDDAP. All buoys will have the same format. However there will be
cases where some variables will have NAs for the entire variable.
}
\examples{
\dontrun{
get_buoy_data(buoyid="42066", vars = NULL)

}

}
