% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Gets list of variables available across all buoys}
\usage{
get_variables()
}
\value{
data frame
}
\description{
Not every buoy contains data for all variables.
}
\examples{
\dontrun{
get_variables()

}

}
