% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbdcbuoy.R
\docType{package}
\name{nbdcbuoy}
\alias{nbdcbuoy}
\alias{nbdcbuoy-package}
\title{nbdcbuoy: Extract buoy data from nbdc}
\description{
Tools to aid in extraction of data from the nbdc website
}
\details{
\itemize{
\item Get_ functions to get data.
\item functions to manipulate the data, aggregate the data
}

To learn more about using \code{nbdcbuoy} click the index below
}
