% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndbcbuoy.R
\docType{package}
\name{ndbcbuoy}
\alias{ndbcbuoy}
\alias{ndbcbuoy-package}
\title{ndbcbuoy: Extract buoy data from ndbc}
\description{
Tools to aid in extraction of data from the nbdc website
}
\details{
\itemize{
\item Get_ functions to get data.
\item functions to manipulate the data, aggregate the data
}

To learn more about using \code{ndbcbuoy} click the index below
}
