% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoyData.R
\docType{data}
\name{buoyData}
\alias{buoyData}
\title{buoyData: List of stations}
\format{A data frame of several elements
\describe{
    \item{ID}{Buoy ID name}
    \item{Y1}{First year of data}
    \item{YN}{Last year of data}
    \item{nYears}{Number of years of data}
}}
\source{
National Data Buoy Center: \url{ndbc.noaa.gov}
}
\usage{
buoyData
}
\description{
A complete list of buoys listed on ndbc website along with buoy specific features such as location and data range.
}
\keyword{datasets}
