% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buoy_data.R
\name{get_buoy_data}
\alias{get_buoy_data}
\title{Gets bouy data from nbdc}
\usage{
get_buoy_data(buoyid = "bzbm3", year, outDir)
}
\arguments{
\item{buoyid}{Character vector. Vector of buoy id names}

\item{year}{Numeric vector. Years of data to pull}

\item{outDir}{Character string. Full path to where output should be written}
}
\description{
Gets bouy data from nbdc
}
